refactor: move the operation_limits module into the limits plugin
