+++ apollo-router/src/plugins/limits/mod.rs
@@ -1,5 +1,6 @@
 mod layer;
 mod limited;
+pub(crate) mod operation_limits;
 pub(crate) mod operation_limits_layer;
 
 use std::error::Error;
